requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.2
django-cors-headers==3.11.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
itsdangerous==2.1.1
itypes==1.2.0
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.0
openapi-codec==1.3.2
Pillow==9.0.1
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
webencodings==0.5.1
Werkzeug==2.0.3
